pyproject.toml
[tool.poetry]
name = "django-chat"
version = "0.1.0"
description = "Chat application with django rest framework"
authors = ["Your Name <you@example.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.0.6"
djangorestframework = "^3.15.1"
djangorestframework-simplejwt = {extras = ["crypto"], version = "^5.3.1"}
django-cors-headers = "^4.3.1"
drf-yasg = {extras = ["validation"], version = "^1.21.7"}
psycopg2 = "^2.9.9"
django-filter = "^24.2"
drf-spectacular = {extras = ["sidecar"], version = "^0.27.2"}


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
